# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ParkSecure API",
    "version": "1.0.0",
    "description": "REST API for the ParkSecure vehicle security platform.\n\n**Authentication:** obtain a JWT via `POST /api/auth/login` and send it as `Authorization: Bearer <token>`.\n\n**Roles:** `super_admin`, `park_admin`, `facility_admin`, `security`, `vehicle_owner`. Staff with an assigned facility are automatically scoped to that facility; vehicle owners only see their own records."
  },
  "servers": [
    {
      "url": "https://parksecure-api.onrender.com"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "Payment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "vehicle_id": {
            "type": "string"
          },
          "facility_id": {
            "type": "string"
          },
          "exit_request_id": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "billing_mode": {
            "type": "string",
            "enum": [
              "free",
              "standard",
              "hourly"
            ]
          },
          "duration_minutes": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "paid",
              "failed",
              "waived"
            ]
          },
          "payment_method": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "plate_number": {
            "type": "string"
          },
          "owner_name": {
            "type": "string"
          },
          "paid_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "facility_id",
          "amount"
        ]
      },
...
      "UserBill": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "created_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updated_date": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "billing_item_id": {
            "type": "string"
          },
          "vehicle_id": {
            "type": "string"
          },
          "facility_id": {
            "type": "string"
          },
          "owner_name": {
            "type": "string"
          },
          "owner_email": {
            "type": "string"
          },
          "plate_number": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "enum": [
              "app_launch",
              "registration",
              "subscription",
              "parking",
              "penalty",
              "maintenance",
              "other"
            ]
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "paid",
              "waived",
              "overdue"
            ]
          },
          "due_date": {
            "type": "string",
            "format": "date-time"
          },
          "billing_month": {
            "type": "number"
          },
          "billing_year": {
            "type": "number"
          },
          "payment_method": {
            "type": "string"
          },
          "paystack_reference": {
            "type": "string"
          },
          "paid_at": {
            "type": "string",
            "format": "date-time"
          },
          "issued_by": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "facility_id",
          "title",
          "amount"
        ]
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/payments/paystack/initialize": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Start a Paystack checkout for a billing record",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "record_type": {
                    "type": "string",
                    "enum": [
                      "payment",
                      "user_bill",
                      "subscription"
                    ]
                  },
                  "record_id": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "required": [
                  "record_type",
                  "record_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authorization_url": {
                      "type": "string"
                    },
                    "reference": {
                      "type": "string"
                    },
                    "access_code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "501": {
            "description": "Paystack not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/payments/paystack/verify": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Verify a Paystack transaction and mark records paid",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reference": {
                    "type": "string"
                  }
                },
                "required": [
                  "reference"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "verified": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/payments/{id}/mark-paid": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Mark a payment as paid manually (staff)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "method": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated payment",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payment"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/payments/{id}/waive": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Waive a payment (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated payment"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/payments/my-balance": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Vehicle owner's outstanding balance",
        "responses": {
          "200": {
            "description": "Balance",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "total_due": {
                      "type": "number"
                    },
                    "bills": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserBill"
                      }
                    },
                    "payments": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Payment"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  }
}